reorganize the api server
