+++ src/bin/bubbaloop.rs
@@ -164,7 +164,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             PipelineMode::Start(pipeline_start_command) => {
                 let response = client
                     .post(format!("http://{}/api/v0/pipeline/start", addr))
-                    .json(&bubbaloop::pipeline::PipelineStartRequest {
+                    .json(&bubbaloop::api::handles::PipelineStartRequest {
                         pipeline_id: pipeline_start_command.id,
                     })
                     .send()
@@ -176,7 +176,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
             PipelineMode::Stop(pipeline_stop_command) => {
                 let response = client
                     .post(format!("http://{}/api/v0/pipeline/stop", addr))
-                    .json(&bubbaloop::pipeline::PipelineStopRequest {
+                    .json(&bubbaloop::api::handles::PipelineStopRequest {
                         pipeline_id: pipeline_stop_command.id,
                     })
                     .send()
